Add: validations to ensure more performance

+++ transcriptor-python/main.py
@@ -8,12 +8,13 @@ from fastapi import FastAPI, UploadFile, File, HTTPException
 
 import webbrowser
 import threading
+import tempfile
 
 def open_browser():
     webbrowser.open_new("http://127.0.0.1:8000/")
 
 # Time to load the swagger UI in the browser, larger time if larger model this is to avoid reloading manually the browser. Else remove threading.Timer
-threading.Timer(5, open_browser).start()
+threading.Timer(2, open_browser).start()
 
 app = FastAPI(docs_url="/")
 
@@ -26,9 +27,9 @@ model = whisper.load_model("large", device = device)
 
 @app.post("/transcribe_from_audio")
 async def transcribe_from_audio(audioFile : UploadFile = File(...)):
+
     # We get the extension of the file to check if it is allowed, and the temp path to read the binary later.
     ext = audioFile.filename.split(".")[-1].lower()
-    temp_path = f"temp_audio.{ext}"
 
     # To avoid injection of .exe files or large files that can crash the service.
     ALLOWED_EXTENSIONS = {"mp3", "wav", "m4a", "webm", "flac", "ogg"}
@@ -39,27 +40,32 @@ async def transcribe_from_audio(audioFile : UploadFile = File(...)):
             detail=f"Unsupported/Not allowed file format '.{ext}'. Please upload one of the following extensions : {', '.join(ALLOWED_EXTENSIONS)}"
         )
 
-    # We write in binary the exact file we are reading, so whisper can read it
-    with open(temp_path, "wb") as audio_file:
-        audio_file.write(await audioFile.read())
+    try:
+        with tempfile.NamedTemporaryFile(suffix=f".{ext}", delete=False) as tmp:
+            tmp.write(await audioFile.read())
+            tmp.flush()
+            temp_path = tmp.name
 
-    # Transcribe the audio file using Whisper
+        # Transcribe the audio file using Whisper
+        start_time = time.time()
+        transcription = model.transcribe(temp_path, language="es")
+        end_time = time.time()
 
-    start_time = time.time()
-    transcription = model.transcribe(temp_path)
-    end_time = time.time()
+        total_time = end_time - start_time
 
-    total_time = end_time - start_time
-    os.remove(temp_path)
+        print(f"Transcription: {transcription['text']}")
+        print(f"Took {round(total_time, 2)} seconds")
 
+        return {
+            "Transcription": transcription["text"],
+            "Language": transcription["language"],
+            "TranscriptionTimeSeconds": round(total_time, 2),
+            "Segments": transcription["segments"]
+        }
 
-
-    return {
-        "Transcription": transcription["text"],
-        "Language": transcription["language"],
-        "TranscriptionTimeSeconds": round(total_time, 2),
-        "Segments": transcription["segments"]
-    }
+    finally:
+        if os.path.exists(temp_path):
+            os.remove(temp_path)
 
 
 """Instalation of the dependencies can be done using the following command : pip install fastapi uvicorn openai-whisper torch"""
